requirements.txt

fastapi==0.111.0
uvicorn==0.30.1

# Database
sqlalchemy==2.0.31
psycopg2-binary==2.9.9

# HTTP & scraping
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
lxml_html_clean==0.4.1
trafilatura==1.7.0

# Config & date handling
pyyaml==6.0.1
python-dateutil==2.9.0.post0

# AI / embeddings (already in project)
openai==1.35.13
tiktoken==0.7.0
numpy==1.26.4
faiss-cpu==1.8.0

# Reliability
tenacity==8.3.0

# Scheduling & background tasks
apscheduler==3.10.4
redis==5.0.4
celery==5.4.0

# Logging
structlog==24.1.0

# Env
python-dotenv==1.0.1
